feat(editinterests): changed a lot in code not much in looks
Re-created the buttons by mapping them, added two proprs so that they can be restyled, added the
click styling, re-ordered the styling. Before it was adjusted for the best design for the buttons to
fit. now, i added them as they were numbered. Please compare the two and if it looks bad i can
revert. all reviews are appreciated :)

#26