feat(event view page): created a page for event view
This page combines EventBanner and EventDescriptionAttendeesList to make a page. This page will need
dynamic routing. Each event will have a different eventID,and that event id will be used to create
the link for each event. For example:localhost:3000/eventID For now, to see how the page would look,
you need to go to localhost:3000/eventview

"re #46"